Move breakpoints to display composable
